Separate ONNX and NCNN nodes
