+++ backend/src/nodes/nodes/ncnn/save_model.py
@@ -1,0 +1,42 @@
+import os
+
+from sanic.log import logger
+
+from ...categories import NCNNCategory
+
+from ...node_base import NodeBase
+from ...node_factory import NodeFactory
+from ...properties.inputs import NcnnModelInput, DirectoryInput, TextInput
+from ...utils.ncnn_model import NcnnModel
+
+
+@NodeFactory.register("chainner:ncnn:save_model")
+class NcnnSaveNode(NodeBase):
+    def __init__(self):
+        super().__init__()
+        self.description = "Save an NCNN model to specified directory. It can also be saved in fp16 mode for smaller file size and faster processing."
+        self.inputs = [
+            NcnnModelInput(),
+            DirectoryInput(has_handle=True),
+            TextInput("Param/Bin Name"),
+        ]
+        self.outputs = []
+
+        self.category = NCNNCategory
+        self.name = "Save Model"
+        self.icon = "MdSave"
+        self.sub = "Input & Output"
+
+        self.side_effects = True
+
+    def run(self, net: NcnnModel, directory: str, name: str) -> bool:
+        full_bin = f"{name}.bin"
+        full_param = f"{name}.param"
+        full_bin_path = os.path.join(directory, full_bin)
+        full_param_path = os.path.join(directory, full_param)
+
+        logger.info(f"Writing NCNN model to paths: {full_bin_path} {full_param_path}")
+        net.write_bin(full_bin_path)
+        net.write_param(full_param_path)
+
+        return True
